requirements.txt
keyboard
pystray
Pillow
wxPython
